Fix build.

+++ helpers.c
@@ -1,6 +1,7 @@
 #include "helpers.h"
 
 #include <limits.h> // for INT_MAX
+#include <string.h>
 #include <stdarg.h> // for va_end, va_start, va_list
 #include <stdio.h>  // for fprintf, stderr, vsnprintf
 #include <stdlib.h> // for malloc, free
@@ -37,7 +38,7 @@ char *d0_asprintf(const char *restrict fmt, ...) {
 char *d0_strndup(const char *s, size_t n) {
   char *out = malloc(n + 1);
   if (out == NULL) {
-    fprintf(stderr, "ERROR: could not allocate %d bytes\n", n + 1);
+    fprintf(stderr, "ERROR: could not allocate %d bytes\n", (int) (n + 1));
     return NULL;
   }
   strncpy(out, s, n);
